Edit is_of_compulsory_school_age file
