+++ policyengine_uk/parameters/gov/dfe/universal_childcare_entitlement/min_age.yaml
@@ -1,4 +1,4 @@
-description: The Department for Education provides universal childcare entitlement in England, specifying the eligible age range.
+description: The Department for Education limits universal childcare entitlement in England to children this age or older.
 metadata:
   type: single_amount
   threshold_unit: year
