+++ policyengine_uk/variables/gov/dfe/universal_childcare_entitlement/universal_childcare_entitlement_eligible.py
@@ -10,11 +10,11 @@ class universal_childcare_entitlement_eligible(Variable):
     def formula(person, period, parameters):
         country = person.household("country", period)
         countries = country.possible_values
-        in_England = country == countries.ENGLAND
+        in_england = country == countries.ENGLAND
 
         age = person("age", period)
         p = parameters(period).gov.dfe.universal_childcare_entitlement
         meets_min_age = age >= p.min_age
         not_compulsory_age = ~person("is_of_compulsory_school_age", period)
 
-        return in_England & meets_min_age & not_compulsory_age
+        return in_england & meets_min_age & not_compulsory_age
